+++ NEWS.md
@@ -1,6 +1,13 @@
+## skewlmm 1.1.0 _(2022-09-27)_
+
+* `rsmsn.clmm` was added to simulate data from SMSN-CLMM with censored responses.
+* `smn.clmm` was added to fit linear mixed models with censored responses.
+* Generic functions such as `print`, `summary`, `plot`, `residuals`, `fitted`, `predict`, and `update` work for objects of class `SMNclmm`.
+* `mahalDist.SMNclmm` computes the squared Mahalanobis distance from a fitted SMN-CLMM. 
+
 ## skewlmm 1.0.0 _(2021-09-17)_
 
-* The main functions `smn.lmm()` and `smnn.lmm()` were changed to pass control options through the `lmmControl` function.
+* The main functions `smn.lmm()` and `smsn.lmm()` were changed to pass control options through the `lmmControl` function.
 * The default estimation method was changed to the DAAREM method, to improve the general performance.
 * A parallel optimization was added using optimParallel.
 * Some options of the diagnostic tools were adjusted.